frontend refactoring
